completely delegate printing VM's state to `VMState`
